refactor: flatten addr_to_f31 into memory
No need for an extra directory, it is legacy from when we had a
id->addr->values separation.

Only move, no logic changes, except from removing `N_BITS_PER_FELT`
which was legacy anyway and unused.
