+++ crates/prover/src/components/memory/component.rs
